requirements_dev.txt
-e .
setuptools==80.9.0
mashumaro==3.17
pdoc==16.0.0
mypy==1.18.2
synthetic_home==4.6.3
hass-client==1.2.0
datasets==4.4.1
pre-commit==4.5.0
ruff==0.14.6

# Dependencies for specific ML integrations
openai==2.8.1
google-generativeai==0.8.5
ollama==0.6.1
anthropic==0.75.0
PyTurboJPEG==1.8.0

# Evaluation dependencies
pytest_homeassistant_custom_component==0.13.272
syrupy>=4.6.1
home_assistant_intents
hassil>=2.0.4
ha-ffmpeg>=3.2.2
pymicro-vad>=1.0.1
mutagen>=1.47.0
pyspeex_noise>=1.0.2
pyrate-limiter==3.9.0
voluptuous-openapi>=0.1.0

# Rely on home-assistant version to pin these
PyYAML
tqdm
